+++ doc/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Made help commands respond in private message to reduce chat spam
 - Added support for Cyberpunk RED game system
 - Added support for Witcher game system
+- Added support for additional wrath dice for wrath and glory
 
 ## [1.3.0] - 2025-07-03